pyproject.toml
[tool.poetry]
name = "boticario"
version = "0.1.0"
description = ""
authors = ["Thiago da Cunha Borges <thcborges@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
pandas = "^1.5.3"
openpyxl = "^3.1.2"
pyyaml = "^6.0"
python-decouple = "^3.8"
pyarrow = "^11.0.0"
spotipy = "^2.22.1"
psycopg2-binary = "^2.9.5"

[tool.poetry.scripts]
boticario = "boticario.cli:main"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.10.3"
pytest = "^7.2.2"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint_check = "blue --check --diff .; isort --check --diff ."
lint = "blue . && isort ."
pre_test = "task lint_check"
test = "pytest -s -x -vv"